requirements.txt
amqp==2.5.2
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-WTF==0.14.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
greenlet==0.4.15
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
libusb1==1.7.1
MarkupSafe==1.1.1
ndeflib==0.3.3
nfcpy==1.0.3
oauth2client==4.1.3
olefile==0.46
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycups==1.9.74
pyDes==2.0.1
pyserial==3.4
pytz==2019.3
qrcode==6.1
redis==3.4.1
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.0
WTForms==2.2.1